Hspec VS tests
