+++ clapi-hs.cabal
@@ -121,7 +121,7 @@ test-suite clapi-hs-test
                      , TypesSpec
                      , UtilSpec
                      , TestValidator
-                     , TestValuespace
+                     , ValuespaceSpec
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N
   default-language:    Haskell2010
